Bug fix
In "screen.h", function "clear_screen()" is define and implement right
inside header file, cause multiple function bug and linker error
Fix bug by move clear_screen() implementation to another source file
